requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
fastapi==0.115.6
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
gotrue==2.12.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
multidict==6.7.0
openai==1.52.2
packaging==25.0
platformdirs==4.4.0
postgrest==0.18.0
propcache==0.4.1
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.6.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.22.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supabase-auth==2.22.0
supabase-functions==2.22.0
supafunc==0.7.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==11.0.3
Werkzeug==3.1.3
yarl==1.22.0
